@@ -623,6 +623,7 @@ class Summary extends React.Component {
 		const {
 			enable_anti_tracking,
 			is_expert,
+			current_theme,
 		} = this.props;
 		const isCondensed = this._isCondensed();
 
@@ -638,6 +639,7 @@ class Summary extends React.Component {
 					isTooltipHeader={is_expert}
 					isTooltipBody={!isCondensed}
 					tooltipPosition={isCondensed ? 'right' : is_expert ? 'top top-right' : 'top'}
+					current_theme={current_theme}
 				/>
 			</div>
 		);
@@ -647,6 +649,7 @@ class Summary extends React.Component {
 		const {
 			enable_ad_block,
 			is_expert,
+			current_theme,
 		} = this.props;
 		const isCondensed = this._isCondensed();
 
@@ -662,6 +665,7 @@ class Summary extends React.Component {
 					isTooltipHeader={is_expert}
 					isTooltipBody={!isCondensed}
 					tooltipPosition={isCondensed ? 'right' : 'top'}
+					current_theme={current_theme}
 				/>
 			</div>
 		);
@@ -671,6 +675,7 @@ class Summary extends React.Component {
 		const {
 			enable_smart_block,
 			is_expert,
+			current_theme,
 		} = this.props;
 		const isCondensed = this._isCondensed();
 
@@ -686,6 +691,7 @@ class Summary extends React.Component {
 					isTooltipHeader={is_expert}
 					isTooltipBody={!isCondensed}
 					tooltipPosition={isCondensed ? 'right' : is_expert ? 'top top-left' : 'top'}
+					current_theme={current_theme}
 				/>
 			</div>
 		);
@@ -740,7 +746,7 @@ class Summary extends React.Component {
 	 * @return {JSX} JSX for rendering the plus upgrade banner or subscriber icon
 	 */
 	_renderPlusUpgradeBannerOrSubscriberIcon() {
-		const { is_expert } = this.props;
+		const { is_expert, current_theme } = this.props;
 
 		const isPlusSubscriber = this._isPlusSubscriber();
 		const upgradeBannerClassNames = ClassNames('UpgradeBanner', {
@@ -752,7 +758,7 @@ class Summary extends React.Component {
 			<div onClick={this.clickUpgradeBannerOrGoldPlusIcon}>
 				{isPlusSubscriber && (
 					<div className="Summary__subscriberBadgeContainer">
-						<div className="SubscriberBadge">
+						<div className={`SubscriberBadge ${current_theme}`}>
 							<ReactSVG src="/app/images/panel/gold-plus-icon.svg" className="gold-plus-icon" />
 						</div>
 					</div>
@@ -779,6 +785,7 @@ class Summary extends React.Component {
 			enable_offers,
 			is_expert,
 			is_expanded,
+			current_theme
 		} = this.props;
 		const { disableBlocking } = this.state;
 		const isCondensed = this._isCondensed();
@@ -787,39 +794,46 @@ class Summary extends React.Component {
 			'Summary--expert': is_expert && !is_expanded,
 			'Summary--condensed': isCondensed,
 		});
+		const foregroundClassNames = ClassNames('Summary__foreground', {
+			active: current_theme === 'palm-theme'
+					|| current_theme === 'leaf-theme',
+		});
 
 		return (
 			<div className={summaryClassNames}>
-				{!isCondensed && disableBlocking && (<NotScanned isSmall={is_expert} />)}
-				{!isCondensed && !disableBlocking && this._renderDonut()}
-				{!isCondensed && !disableBlocking && this._renderPageHostReadout()}
+				<div className="Summary__opacityOverlay" />
+				<div className={foregroundClassNames}>
+					{!isCondensed && disableBlocking && (<NotScanned isSmall={is_expert} />)}
+					{!isCondensed && !disableBlocking && this._renderDonut()}
+					{!isCondensed && !disableBlocking && this._renderPageHostReadout()}
 
-				{isCondensed && !disableBlocking && this._renderTotalTrackersFound()}
+					{isCondensed && !disableBlocking && this._renderTotalTrackersFound()}
 
-				<div className="Summary__pageStatsContainer">
-					{!disableBlocking && this._renderTotalTrackersBlocked()}
-					{!disableBlocking && this._renderTotalRequestsModified()}
-					{!disableBlocking && this._renderPageLoadTime()}
+					<div className="Summary__pageStatsContainer">
+						{!disableBlocking && this._renderTotalTrackersBlocked()}
+						{!disableBlocking && this._renderTotalRequestsModified()}
+						{!disableBlocking && this._renderPageLoadTime()}
+					</div>
+
+					{isCondensed && disableBlocking && (
+						<div className="Summary__spaceTaker" />
+					)}
+
+					<div className="Summary__ghosteryFeaturesContainer">
+						{this._renderGhosteryFeature('trust')}
+						{this._renderGhosteryFeature('restrict', 'Summary__ghosteryFeatureContainer--middle')}
+						{this._renderPauseButton()}
+					</div>
+					<div className="Summary__cliqzFeaturesContainer">
+						{this._renderCliqzAntiTracking()}
+						{this._renderCliqzAdBlock()}
+						{this._renderCliqzSmartBlock()}
+					</div>
+					{this._renderStatsNavicon()}
+					{enable_offers && this._renderRewardsNavicon()}
+
+					{!isCondensed && this._renderPlusUpgradeBannerOrSubscriberIcon()}
 				</div>
-
-				{isCondensed && disableBlocking && (
-					<div className="Summary__spaceTaker" />
-				)}
-
-				<div className="Summary__ghosteryFeaturesContainer">
-					{this._renderGhosteryFeature('trust')}
-					{this._renderGhosteryFeature('restrict', 'Summary__ghosteryFeatureContainer--middle')}
-					{this._renderPauseButton()}
-				</div>
-				<div className="Summary__cliqzFeaturesContainer">
-					{this._renderCliqzAntiTracking()}
-					{this._renderCliqzAdBlock()}
-					{this._renderCliqzSmartBlock()}
-				</div>
-				{this._renderStatsNavicon()}
-				{enable_offers && this._renderRewardsNavicon()}
-
-				{!isCondensed && this._renderPlusUpgradeBannerOrSubscriberIcon()}
 			</div>
 		);
 	}
